Update CentOS Stream 9 info and container image
CentOS Stream is not a rolling distribution.  It was originally
advertised as one to describe how it rolls between minor versions.
However, it still has major versions and EOL dates, which by the common
definition makes it not a rolling distribution.  Due the confusion, the
project stopping using that term to describe the distribution.

The EOL date for CentOS Stream 9 will match the date of the end of the
RHEL 9 "Full Support" phase, which will be 5 years after the RHEL 9
launch, which will be approximately 3 years after the RHEL 8 launch.

This also switches the container image from the stream9-development tag
to the stream9 tag, which includes the ppc64le and s390x architectures.

https://en.wikipedia.org/wiki/Rolling_release
https://centos.org/centos-stream/
https://access.redhat.com/support/policy/updates/errata#Full_Support_Phase